requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.20
pytest